fix: rework the implicit return protocol and fix several bugs with its usage (#590)
Fixes #589

Switch expressions and conditional expressions can both be patched as IIFEs, in
which case they get to decide how to handle sub-expressions in an implicit
return position. However, neither of these classes implemented the methods to
declare that they are an implicit return patcher, so the code was using the next
ancestor instead when searching for the implicit return patcher. That meant that
a switch or conditional IIFE within a loop IIFE would include `.push` lines in
the inner IIFE when it shouldn't.

Another issue was that loop patchers were declaring themselves as an implicit
return patcher even when they were patched as a normal expression and not an
IIFE. I'm not aware of any correctness issues this caused, but it seems good to
fix.

To fix, I changed conditional, switch, and loop patchers to all properly declare
when they should decide the behavior of subexpressions in an implicit return
position. I also reworked the protocol to hopefully be a little simpler; any
class just declares when it's an implicit return patcher, and there are a few
other hooks that LoopPatcher uses to provide custom behavior.