fix: rework constructor patching code to better handle bound methods
Fixes #596
Fixes #597

When patching constructors when there are bound methods, we need to insert code
after the super statement, if there is one. This is inherently a bit fragile,
since magic-string works best when all code is patched in order. Previously, we
were patching the method body, then inserting the bindings, but that could cause
the new statements to be inserted after the `}`. To fix this, I made the code
smarter about patching order: it patches the function either before or after the
inserted statements to avoid interfering with the start or end. This solution
feels a bit hacky, but at least for now seems nicer than trying to actually
force everything to patch in order.

Also, the code was assuming that the body was non-null, so I added a special
case with a null body where we overwrite the method body with the proper
contents.