feat: implement strictly correct default params and enable them by default
Fixes #125

Now, instead of converting CS default params to JS default params, the default
behavior is to emit assignments behind a null/undefined check. This is more true
to CS semantics: in CS, defaults are used when the argument is null or
undefined, while in JS, they are only used when the argument is undefined.

There was already similar code for handling `this` assignment in parameters, so
I based default parameter code off of that. There were various special cases to
consider:
* In a function like `(@a=b) -> c`, we need to generate code for both the
  default assignment and the `this` assignment, and the variable names need to
  match up.
* In a function like `({a}={}) -> b`, we need to do the null check before the
  destructure since the null check needs to be on a regular variable, so we need
  to introduce a new variable.
* If the destructure is multiline, moving it inside the function might end up
  with code with bad indentation, so adjust the indentation to avoid this.

After handling these cases, decaffeinate can handle all files in my codebase
without crashing, and it doesn't introduce any obvious new correctness issues.

This commit also adds a `--loose-default-params` option to disable this
behavior and instead output the old code that was cleaner but less correct.