refactor: consolidate patcher constructor args into a PatcherContext type
I want the decaffeinate options to be accessible from any patcher, but currently
we have exactly three constructor args that we forward all over the place. To
make it easier to add new constructor args, I made a new type for those three
and changed everything to use that.

This was done with some simple find-and-replace operations, but given how good
the test coverage is and how explosive any mistakes would be, it seems like it
should be fine to make a widespread change like this without needing to be too
careful.