fix: improve performance when searching tokens between source indices
Fixes #612

Previously, whenever we searched between two nodes or two source indices, we
would do a scan through the whole token list. Now, we limit it to just the
region that we're interested in.

This also caused an issue when searching for an open-paren after a bare `super`,
since some nodes are virtual and have undefined token indices, so I added a
special case there so we just say that the bare `super` case never counts as an
implicit call.

With this and the other changes, the file in #612 goes down to about 10 seconds
to convert, whereas before it took over 10 minutes. Also, converting my whole
codebase got a little faster (about 20% less time to convert).