refactor: rework makeRepeatable to be more integrated with patching
Progress toward #592

Previously, `makeRepeatable` was a function that you would call around the time
of `patch` (sometimes before, sometimes after) that would sometimes do some
magic-string operations to extract the expression into a variable. However,
this had some problems:
* To be friendly with magic-string, all code should be patched in order, meaning
  that some of the `makeRepeatable` code should run before patching the node and
  some should run after, rather than them being separate operations.
* Because there was no obvious relation between patching and `makeRepeatable`,
  the `isRepeatable` case couldn't assume that the code was already patched and
  just returned the old CoffeeScript code, which only worked in simple cases
  like identifiers. Even if the expression was already patched, getting the
  accurate code after patching is hard; getting the code while patching, like in
  `patchAndGetCode`, is more reasonable.
* `ThisPatcher` needed a special case to work around some of these issues
  (particularly related to the fact that sometimes the CoffeeScript code is just
  `@`).

The new way of thinking about `makeRepeatable` is similar to
`setRequiresExpression` and `setImplicitlyReturns`: you call
`setRequiresRepeatableExpression` to inform the patcher that whenever it runs
its patching code, it must do so in a way that the expression is repeatable, and
the code to access the value again is then accessible as `getRepeatCode`. The
new `patchRepeatable` method is a new external-facing method that roughly
replaces `makeRepeatable` and also does patching at the same time, so most cases
could be moved to just use that.

While I'm generally not a big fan of the stateful style of doing things, it
ended up being the most reasonable approach I could think of, especially when
dealing with patchers like `ChainedComparisonOpPatcher` that want to run the
regular patching code with some modifications, then access the repeat code for
expressions relatively deep within the AST.