fix: update esnext and fix a scoping issue introduced by the bug fix (#629)
decaffeinate's code for handling IIFE-style loop expressions would insert a
`var` declaration for a temporary variable to receive the value for the current
iteration. But a `var` declaration is actually wrong, because in some cases it
would end up with the previous loop iteration's value, so we should generate
`let` instead.