fix: only use assignment-style repeatable expressions when necessary (#632)
Fixes #592

Patching for soak operations is a bit fragile because the `__guard__(` and the
close-paren are inserted in code outside of the soak operation itself. This is
made more complicated with repeatable expressions, since both want to surround
the code in similar ways.

In member access expressions (including soaked member access), the code to make
the expression repeatable needed more complexity to make it so the repeated
result is still a valid assignee, but this complexity was only needed in two
assignment cases. To limit problems caused by this approach, and make the code a
little nicer in some cases, the behavior is now opt-in: when making an
expression repeatable, you can pass an `isForAssignment` flag to indicate that
the patching needs to be done in a way that the repeated code is valid to assign
to, not just valid to use an expression.

The problem from #592 doesn't come up for assignments because the soak container
(as evaluated by `findSoakContainer`) includes the whole assignment expression,
so code isn't being inserted in a conflicting way. So this doesn't feel like an
excellent solution, but it works in all cases that I know. Maybe in the future,
`__guard__` patching will need to be changed to always patch in order (or
replaced with another approach), but this works for now.

Also fix an issue where the fix for #512 needed to be applied to logical
compound assignment operators as well.