feat: change loops and loop expressions to be strictly correct by default (#633)
Fixes #523

Decaffeinate now uses `Array.from` in both JS for-of loops and when using
`map`/`filter`, which handles the case where the loop target is an array-like
object (with `length` and numerical indices) instead of a real array (which is
iterable and has `map` and `filter`).

I made two separate prefs for these, since it looks like there are many more
array-like objects in the wild without `map` and `filter` (or where the
semantics are different, e.g. jQuery collections) than there are array-like
objects that are not iterable. But both seem to exist, so to stay on the safe
side, I'm wrapping in `Array.from` by default in both cases. It may make sense
to enable `--loose-for-of` by default in the future if it's deemed safe enough.