fix: avoid out-of-order patching in manually bound functions (#634)
Fixes #579

The function patching code was in some cases inserting a `}` at `innerEnd` and
then inserting `.bind(this)` at `contentEnd`. Usually those are the same (and
this is a rare case anyway, so it didn't come up much), but in the example in #579,
`innerEnd` was after the newline and `contentEnd` was before, so `.bind(this)`
was being added to early. To fix, I changed the `.bind(this)` code to go at
`innerEnd`, which generally follows the rule that patching should go in order.