feat: implement strictly-correct `in` patching by default
Fixes #636

There was already code to optionally surround in parens, so we just needed a new
case to optionally surround in `Array.from(` and `)`. As with other `Array.from`
usages, I added a flag to disable this behavior if people are confident that
they're only working with arrays or otherwise objects that have `includes`.

Also go through some old tests and fix their names to better reflect what's
actually done.