fix: don't use `map` when a loop assignee can't safely be a parameter
Fixes #635

In some cases, we would change a variable to a parameter when converting an
expression-style loop, and this caused other usages to be invalid. In other
cases, there could be loop assignees like `@a` that aren't valid parameters (at
least not without additional work). To avoid these cases, we can add an
additional check when deciding patching style, where we only proceed if all
assignees are just identifiers and if those identifiers are only used within the
loop itself.