fix: rework many of the details of range iteration
Fixes #640
Fixes #642

There were a number of details that weren't quite right in loop iteration, so I
reworked a lot of it:
* It's important to *always* use a C-style loop when iterating through a range,
  since the final value of the loop assignee is meaningful in some cases. For
  example, code might loop through `[0...length]` and then check if `i == length`
  at the end to see if a break ever happened. This meant that I had to implement
  the case where we need to decide at runtime which direction to iterate.
* Changing the end or the step of the iteration while the loop is in progress
  should not affect the loop, so we need to save those values to variables at
  the start. In the future, we could potentially do some fancy static analysis
  to see if these variables could be assigned within the loop and skip the
  initial assignment in that case.
* When iterating through a normal array-like object with a variable step, we
  need to dynamically check whether the step is positive or negative to know if
  we should iterate forward or backward.

I tried to avoid variables when possible, e.g. if the end is a constant, then we
don't need to assign it at the start of the loop. There are probably even more
ways to avoid extra variable usages in specialied cases, but those can come
later.