feat: implement non-terminal rest in array destructures
Progress toward #107

We already had code to handle expansion nodes, so non-terminal rest is almost
the same, we just need to produce one more assignment that slices the middle
part of the array.

I also reworked some of the code to use `patchAndGetCode` when possible and to
patch code in order as much as we can.

The plan is to implement non-terminal rest and expansion within args by moving
the trailing parameters to an array destructure in the normalize step, so we'll
only need one code path that needs to think about the details of generating this
type of code.