fix: don't patch past the end of the expression in insertStatementsAtIndex (#652)
Fixes #648

Various code paths use `getBoundingPatcher` to determine the latest index that
we can safely patch code, which is especially important when a line ends with
close-parens and/or close-braces that close outer nodes. The
`insertStatementsAtIndex` method can be used to add a line to the end of a
block, but it wasn't doing this check, so in some cases it would incorrectly
insert the new line after a close paren that's supposed to be after the block.