fix: change EqualityPatcher to respect the needsParens flag (#657)
Fixes #656

`EqualityPatcher` already extended `BinaryOpPatcher`, and `BinaryOpPatcher`
already exposes a way to customize operator patching, so we can just use that
and we'll also pull in the needsParens behavior (importantly, telling child
assignment expressions to add parens). This adds parens in a few cases where
they're unnecessary, which is arguably a little uglier, but eslint has an
auto-fixer for this with the no-extra-parens rule, so that probably handles
typical use cases if people case.