fix: wrap loop components in parens in the normalize stage if necessary
Fixes #663

We want to avoid an unparenthesized `then` token in a `while` condition, since
that will confuse the CoffeeScript parser. Ideally, maybe we'd tokenize the new
condition that results from the normalize step, or detect this case in some
other fancy way, but an easy approach that works is to just see if `then` is a
substring of the expression code, and defensively wrap in parens if so.

The same issue comes up for `for` as well, so take a similar approach there.

These tests also exposed some issues with `for` and `while` patching in general,
mainly that we weren't setting fields to be expressions when we should have, and
we were using too lenient of a range when finding the `then` token after a loop.