fix: patch the step before rearranging loops in the normalize stage (#662)
Fixes #661

When rearranging a node, we should patch the children first, then work with the
patched results. For for-in loops in particular, we were patching the step after
everything else, so patching the step could cause magic-string problems. An easy
fix is to just reorder it so we patch the step first.