fix: change the default negate function to insert at contentStart (#679)
Fixes #678

Previously, the default behavior was to insert at `outerStart`, which would
usually crash if the expression was surrounded by parens, since `outerStart` is
outside the editing bounds in that case. We can't rely on having parens anyway,
so any subclasses should (and already do) add extra parens if necessary.