feat: overhaul assignment patching to allow nested destructuring
Fixes #268

CoffeeScript assignment has some differences from JavaScript assignment in a few
ways:
* CoffeeScript allows the expansion operator (`...`).
* CoffeeScript allows intermediate rest assignments.
* CoffeeScript allows splice operations at any point in an assignment.
* CoffeeScript destructure defaults will fall back to the default if the value
  is `null` or `undefined`, unlike JS, which only falls back if the value is
  `undefined`.
* CoffeeScript array destructuring happens by accessing `arr[0]`, etc, while
  JavaScript destructuring uses the iterable protocol.

Assignment patching now has two code paths: the "normal" patching code path, and
the "expanded" code path. We use the normal code path when possible, which
translates the assignee in a straightforward way, but if any part of the
assignee can't be normally tranformed, we use the expanded code path as a
fallback. The expanded code path generally rewrites the assignment as a sequence
of intermediate assignments. Both code paths wrap the expression in `Array.from`
when the assignee is an array destructure.

For the expanded code path, we give up on attempting to do precise patching due
to the complex recursive nature of the problem. Instead, we generate a big
string of code and replace everything with that string. Since this can destroy
formatting, we prefer the normal code path when possible rather than making it a
special case of the expanded code path.

Note that this does not yet handle arbitrary parameters and loop assignees. The
plan is to change those to assignment statements in the normalize stage in a
later commit.