feat: implement splice operations
Fixes #668

This is heavily based off of the slice patching code, with some modifications.
It's also a little strange in that the slice patcher patches only part of the
slice call and the assign op patcher patches the rest, but that seemed like the
cleanest way to do it with the way things are structured now.