feat: more intelligently handle invalid constructors
Fixes #591
Fixes #726

A major difference between CoffeeScript and JavaScript class constructors is
that JavaScript class constructors require you to call `super` and require it
to be called before the first use of `this` (when the class is a subclass).
This has caused two significant issues in decaffeinate:
* CS code that uses `this` before `super` or omits `super` will convert to code
  that is rejected by babel. This happened for 11 files in my codebase.
* Subclasses that use bound methods will put the binding code after the `super`
  call, while the CoffeeScript semantics are to put it before the `super` call.
  This caused (at least) two correctness issues in my codebase, only one of
  which was caught the tests. (Both of these were backbone subclasses where the
  `initialize` method gets called before the super call ends.)

Since there doesn't seem to be a great answer to this problem, this commit
provides three options for how to handle it:
* Give an error rather than generating incorrect code. This is the default
  behavior. If a constructor uses `this` before `super` or omits `super`, or if
  a subclass uses a bound method, then decaffeinate gives an error for the file
  explaining the issue and the different options. This flags the issue for
  people not aware of it, and gives people a chance to change their CoffeeScript
  code to not have these issues (e.g. rewriting constructors and/or changing
  bound methods to explicit bindings).
* `--allow-invalid-constructors` generates JS code even if `this` is used before
  `super` in the constructor, and it generates code that would be correct if JS
  allowed it (method bindings at the start of the constructor). Technically, this
  code is syntactically valid, but will crash at runtime in a JS interpreter and
  will be rejected by babel. The intention here is that it may be useful to
  convert the code anyway and fix the invalid constructors later, e.g. moving
  the binding lines after doing a sanity check.
* `--enable-babel-constructor-workaround` generates the same invalid code, but
  also puts a code snippet at the top of each constructor that tricks babel into
  allowing `this` before `super` in constructors (even at runtime). This works
  because babel doesn't use fully-compliant JS classes, and it's certainly a
  hack that might break due to a change in babel, but as far as I can tell, it's
  completely correct as long as you're using babel. This is especially useful as
  a way to automatically convert a project and run all of the tests, but
  probably shouldn't be used in production or at least should be removed as soon
  as possible if it is.

Notably, there is no longer a way to put method bindings after the super call.
This both makes the implementation simpler and is more in line with "either
generate correct code or give an error". There could be an option for this in
the future, but really, I think it's best for people to manually check that
reordering the constructor is ok.

Another possible future option is to generate classes the same way that
CoffeeScript does, or do some other less hacky trick to allow `this` before
`super`. But JS classes fundamentally don't allow this, and CS classes can't
extend JS classes, so it seems like the only 100% safe way to do this is to
NEVER generate JS classes, which seems like it would be more work and defeat a
lot of the purpose of decaffeinate.