fix: un-break CLI by changing detect-indent import to a form recognized by rollup
This fixes a consistent crash caused by #735 where any non-trivial file would
fail due to `detectIndent` not being defined. From what I can tell, rollup can't
handle the `import detectIndent = require('detect-indent');` line and was just
ignoring it, so `detect-indent` was never pulled in and everything broke when it
tried to use that method.

A little ironic that this would have been caught by tests were it not for #712.
But probably it's still fine to leave the tests as testing the source code and
not the dist build.