fix: make edit bounds checking within an array more lenient
Fixes #740

I think fixes a regression from #719.

The assertion that we edit with bounds should be a little more lenient than the
code that limits writing to the bounding patcher. We already had a case for that
for function parameters, so add one for array elements as well.

The resulting code in the test is a bit ugly, but at least it doesn't crash.