fix: correctly handle splice operations on an individual element (#752)
Fixes #751

I mistakenly thought `Array.from` would do what I want, but it actually just
happens to work for single-character strings because `Array.from('a')` is
`['a']`. Instead, we should use `concat` like CoffeeScript does (which has its
own rules about when to keep something an array and when to treat the arg as a
single element). I also made the tests a little more interesting to expose this
case.