feat: handle complex function parameters
Fixes #756

Currently, this is implemented by taking the first parameter that can't be
converted in a straightforward way and moving everything from there later to a
rest param in the normalize step. This lets us reuse code for complex
assignments in the main stage. This certainly has some room for improvement,
since in some cases it would be nice to do complex assignments without using
rest params, but code like this is probably rare enough that it's not worth
worrying about too much.