fix: defer the soak close-paren until after the soak container is patched
Fixes #704

In some cases, a soak operation could place a close-paren at the end of the soak
container before the normal suffix for the soak container node. To avoid this, I
added a way to defer a code snippet to be inserted at the end of a node once it
is finished patching.

It's almost the case that you can just apply the suffix code at the end of the
`patch` method, except that the code for handling repeatable expressions needs
to sometimes include the suffix code in its repeat code. So code handling
repeatable expressions needs to be aware of the deferred suffix mechanism, but
hopefully that will stay fairly small-scoped.