fix: better handle comment-only blocks in conditionals
Fixes #774

We used to remove from the end of the consequent to the start of the else in the
expression case, but that can destroy comments, so be more conservative now.
Also, if we have a block with 0 statements in an expression context, we should
use `undefined` instead of having no contents at all.