fix: wrap yield expressions in parens when necessary
Fixes #776

For many expression cases, JavaScript requires parens around yield expressions,
so we need to add them unless we know it's safe not to. yield in a statement
context, yield in a return, and yield on the RHS of an assignment all do not
require parens, so we skip the parens in those cases.