fix: avoid placing close-parens with too much indentation (#781)
Fixes #777

This feels like a bit of a hack, unfortunately, but it's the only thing I could
think of the works in all cases. Rather than placing the close-paren just before
the following dot, we move the dot to be just after where the close-paren would
go, which generally might join the two lines, and might end up destroying an
inline comment, but handles all known cases without crashing.