feat: handle class expressions for complex classes (#799)
Fixes #625

The general strategy here is to define an `__initClass__` helper in the
normalize stage which calls the `initClass` static method. This is uglier than
the previous approach of adding an `initClass` statement after the class
definition, so we dynamically decide which of the two to use, preferring the
statement approach. Since the normalize stage does not properly communicate
whether patching is done as an expression, we just look at the AST and only
allow the `initClass` statement if we're a child of a block and not in an
implicit return position.

Also, the existing code for handling helpers was all for the main stage, so I
factored it into a `SharedProgramPatcher` that also works in the normalize
stage.

This almost fixes #544, but there's a remaining issue in that example because
commas can split class properties but not statements in CoffeeScript.

There's still a little more to do to make complex classes 100% correct, but I
think this should handle most cases.