feat: avoid the guard helper in some soaked member accesses
Progress toward #801
Progress toward #336

Now, with a simple soak expression like `a?.b`, it will be expanded into a
ternary rather than using the `__guard__` function. This allows us to handle the
case where the variable is unbound (in which case the code should return
`undefined` rather than throwing an exception). More specifically, we use the
conditional form of soak operations

As described in #336, I separate the expression and statement cases and prefer
to wrap in an `if` if possible.

This only handles the case for normal soaked member access; a later commit will
expand it to soaked dynamic member access and soaked function application.

Also fix a subtle bug in repeatable expression patching where a paren could be
misplaced.