fix: don't put code in comments when removing trailing empty returns
Fixes #816

The previous logic was to remove trailing empty returns by deleting from the
start of the previous newline to the end of the return, but if there was code
after the `return` on the same line and the previous line had a comment, it
could cause that code to be pulled into the comment.

To fix, I rewrote the logic so it only removes the line if the `return` is on a
line by itself. If not, it does the safe thing of just removing the `return`.

This makes some generated code cleaner, but ends up with an unnecessary blank
line in the `initClass` method for complex class expressions. I couldn't find a
clean solution to avoid this, so it's a style regression in this commit,
favoring correctness in other cases. I think probably the right solution there
is to rethink that code transform so that a line with only `)` doesn't produce
a blank line.