feat: avoid the guard helper in some additional soak cases (#833)
Fixes #801
Progress toward #336

We now use a conditional for soaked dynamic member access and soaked function
calls as long as the expression is repeatable and does not contain any soak
operations.

This is pretty much the same strategy as for soaked member accesses. For
function calls, the code naturally becomes method call syntax when necessary, so
we don't need a special case for methods like with the `__guardMethod__` stuff.

There's some duplicated code, but a lot of helper methods have been shared and
there are enough differences that it seems reasonable to have the duplication
for now and try to consolidate later if it becomes a bigger issue.