feat: allow external constructors in classes
Fixes #830

We handle external constructors in the normalize step, since it generates
another line in `initClass` and that way we don't need to worry about handling
bound methods and `super` issues at the same time.

Note that this does not yet handle external constructors in experession-style
classes, but should should start working once #835 is addressed.