fix: update tests, stop comment patching from overwriting other code
Fixes #838

A decaffeinate-parser change made some AST node bounds different (generally for
the better), which broke a test. This also exposed a regression where
decaffeinate could `insertLeft` at the end of a block comment, and then later
when the comment was transformed, it would overwrite that new code. Instead, we
now transform comments first.

Also add a skipped test for a case I discovered where we destroy a comment.