fix: change negated patching to patch in order
Fixes #840

Rather than immediately surrounding an expression in parens, we should set a
flag on the expression so it gets an open-paren at the start of patching and a
close-paren at the end. There still could potentially be some conflicts between
deferred patching and patching at the end of an expression, but it seems to not
come up at least for this issue.