fix: don't crash on a simple super call within an anonymous class
Fixes #843

For most super calls, we can convert them to JS super without needing a
reference to the enclosing class. However, in more advanced cases, we'll need a
reference to effectivelly call super ourselves. CoffeeScript generates code
where even anonymous classes can be referenced, so we may need to do that at
some point in the future. For now, though, we should just convert to JS super
rather than failing if we can't find a name for the enclosing class and it's a
simple case that can become JS super.