fix: allow computed keys for class methods
Fixes #847

This simplifies the computed key logic to only happen in one place, and ensures
that at least for the normal key case, we actually end up calling patch on the
key. I think some additional work will be necessary to do the equivalent for
static methods, but this should handle the common case.

This also changes the behavior so method name strings aren't surrounded by
brackets, but it turns out that wasn't necessary.