fix: allow duplicate class declarations
Fixes #855

We now track class declarations in our scope, and only use the normal class
syntax if the class being patched is also the declaration site (the first
assignment seen by the scope).

This also lets us remove a special case in the `initClass` patching in the
normalize step. We don't need to explicitly create an assignment anymore,
because one is created automatically due to a variable with the same name being
declared above the class declaration.