fix: properly set the class name for complex class declarations (#862)
Fixes #861

This is mostly a revert of #854, although I also added some additional logic to
check if the generated name is a reserved JS word (copying CoffeeScript's set of
names) and add an underscore if so. I also added some tests to ensure that the
behavior matches CoffeeScript for these cases.