fix: properly specify parens under unary operators (#873)
Fixes #872

The previous logic for unary operators was to propagate the `needsParens`
patching flag to the child, but that actually isn't valid. Even if the unary
operator is in a context where it doesn't need additional parens (e.g. a
conditional), its operand might be a binary operator that needs parens, so just
pass true for `needsParens`.