fix: avoid bounds error when patching an implicit object literal (#877)
Fixes #689

There's a lot of logic to find a nice-looking place for the open-brace, but
generally it might be in a safe place that's before our bounds. But that means
that other patchers, like the exists op patcher, might have already written
there. To avoid this, we now call `adjustBoundsToInclude` to effectively
"reserve" that area. This required a little bit of extra care so that we only
actually adjust the bounds when we're being patched as an expression, and
ExistsOpPatcher needs to patch its child as a statement when necessary.