fix: properly patch code around for-of loop targets (#885)
Fixes #884

In some cases, some code could end up left over on the left side after patching
the target, due to using magic-string's `appendLeft`. To avoid this, we can
always do an overwrite operation for the target and its surrounding code when we
know that we're extracting the target into a variable.