fix: improve the check for an arrow function body starting with an object
Fixes #889

If a JS arrow function has an expression body starting with a `{`, then it needs
to be wrapped in parens. Ideally, we'd check the code itself, but due to patch
ordering details it's best to determine it just from the AST so that the
open-paren is patched before the body itself. So, instead, we can do a traversal
and see if there is any object initializer in the body that starts at the start
of the expression.

Also change one of the cases to patch in order instead of using
`surroundInParens`.