fix: remove the Unicode BOM if necessary at the start of the code (#892)
Fixes #891

Unicode byte order mark characters normally aren't a problem for us, except that
CoffeeScript has a special case to remove them at the start, so it gives all
source indexes as if the BOM was removed. To avoid running into issues there, we
can just do the same. As with newline formats, this is just handled at the
decaffeinate layer and coffee-lex and decaffeinate-parser can assume the simpler
format (although they could also be made more robust later if it's useful).