fix: don't patch out of bounds when placing close-parens
Fixes #744

In some cases, the computed `innerEnd` (which usually defines the editing
bounds) can be strictly before the following close-paren, so trying to patch at
the start of the following close-paren will give an error. In other cases, we
need to patch after `innerEnd` and the editing bounds are extended. To handle
both cases, we allow patching up to the end of the editing bounds, but not
further.