fix: properly handle nested implicit object initializers (#899)
Fixes #898

One issue was that the patching bounds were not properly accounting for object
parents, so add that case. That fixed the correctness issues, but resulted in
ugly code. To avoid inserting a newline unnecessarily, we only add a newline
before the close-curly if we are multiline or if our enclosing object
initializer is multiline.