fix: surround postfix if/while/for in parens when necessary (#897)
Fixes #896

Whenever an if/while/for is in a non-final position in a comma-separated list
and is followed by a comma token, rearranging it will confuse the parser because
the comma will be seen as part of the body. In these cases (which we can detect
by seeing if the node has a comma after it), we can wrap the node in parens to
avoid the ambiguity.