fix: don't unsafely negate operators like `>` by default (#903)
Fixes #902

For both normal binary operators and chained comparison operators, we need to be
careful about rewriting `!(a < b)` as `a >= b`, since the rewrite isn't correct
for cases like `NaN`. Now, we just do the simpler negation by default and only
do the nicer-looking one with an opt-in flag.