fix: extract catch assignees to assignments when necessary
Fixes #585

Catch assignees have different scoping rules in CoffeeScript and JavaScript, so
the simple way to solve that is to generate an assignment at the start of the
catch block so the variable is function-scoped. This is always correct, but
usually ugly, so we only do it in the rare cases where it's necessary:
* when the variable is actally used outside of the catch block.
* when the variable is a complex assignment instead of just an identifier, e.g.
  a destructure with an expansion node. Even though JS does support some forms
  of exception destructuring, for simplicity, we always extract non-identifiers.

From a code standpoint, this is similar to how normalize ForPatcher moves the
assignee to the start of the body when necessary.

Also fix scope assignment traversal to account for catch assignees.

Also fix an error formatting crash that I ran into in an intermediate test
failure.